Factor out parser\anystate.rs from parser\shared.rs
