+++ nanvm-lib/src/parser/mod.rs
@@ -2,6 +2,7 @@
 // It uses src/ast APIs.
 #![allow(clippy::module_inception)]
 pub mod analyzer;
+pub mod any_state;
 pub mod parser;
 pub mod path;
 pub mod shared;
