+++ nanvm-lib/src/parser/parser.rs
@@ -1,10 +1,10 @@
-use super::shared::{
-    AnyResult, AnyState, ConstState, JsonState, ParseError, ParseResult, ParsingStatus, RootState,
-    RootStatus,
-};
 use super::{
+    any_state::{AnyResult, AnyState},
     path::{concat, split},
-    shared::{JsonElement, ModuleCache},
+    shared::{
+        ConstState, JsonElement, JsonState, ModuleCache, ParseError, ParseResult, ParsingStatus,
+        RootState, RootStatus,
+    },
 };
 use crate::{
     common::default::default,
